"ads2_m6_a1_1: 0/12 testcases passed. Cannot run hidden testcases as Server is unreachablecheck_style score: 1/1 "

+++ ADS-2-assignments/m6/ADS-2Week-1/Solution.java
@@ -21,7 +21,7 @@ class PageRank {
         double prval = pR[vertice];
 		for(int j = 0; j < 1000; j++) {
 			for (int i : adJ) {
-				prval += pR[vertice] / dgraph.outdegree(i);
+				prval += (double) (pR[vertice] / dgraph.outdegree(i));
 			}
 			pR[vertice] = prval;
 		}
